requirements.txt
# This file was manually generated.
pandas
pytrakt
rich
pyyaml
cronsim
pydantic
selenium
webdriver-manager
